# src/main/resources/v1/data-catalog-openapi.yaml
openapi: 3.0.1
info:
  title: Data Catalog
  description: API for Data Catalog
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: PDLEAMZEUG@pdl.internal.ericsson.com
  license:
    name: COPYRIGHT Ericsson "201"9
    url: https://www.ericsson.com/en/legal
  version: 1.0.0
servers:
- url: //localhost:8080/catalog
...
paths:
...
  /v1/notification-topic:
    get:
      tags:
      - /v1/notification-topic
      summary: get all NotificationTopic details or filter by QueryParams
      operationId: getAllNotificationTopicsUsingGET
      parameters:
      - name: messageBusId
        in: query
        description: messageBusId related with NotificationTopic
        schema:
          type: string
      - name: name
        in: query
        description: name of the NotificationTopic
        schema:
          type: string
      responses:
        "200":
          description: get NotificationTopic details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationTopicDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: If NotificationTopic details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
    post:
      tags:
      - /v1/notification-topic
      summary: create NotificationTopic details
      operationId: registerNotificationTopicUsingPOST
      requestBody:
        description: NotificationTopic
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTopicDto'
        required: true
      responses:
        "201":
          description: created NotificationTopic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTopicDto'
        "400":
          description: Bad Request in values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: NotificationTopic details exists already
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
      x-codegen-request-body-name: NotificationTopic model
  /v1/notification-topic/{id}:
    get:
      tags:
      - /v1/notification-topic
      summary: get NotificationTopic by Id
      operationId: getNotificationTopicByIdUsingGET
      parameters:
      - name: id
        in: path
        description: id uniquely identifies NotificationTopic
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get NotificationTopic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTopicDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: If NotificationTopic details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
components:
  schemas:
...
    CatalogResponse:
      title: CatalogResponse
      type: object
      properties:
        message:
          type: string
        path:
          type: string
        timeStamp:
          type: string
          format: date-time
...
    MessageBusDto:
      title: MessageBusDto
      type: object
      properties:
        accessEndpoints:
          type: array
          items:
            type: string
        clusterName:
          type: string
        id:
          type: integer
          format: int32
        messageDataTopicIds:
          type: array
          items:
            type: integer
            format: int32
        messageStatusTopicIds:
          type: array
          items:
            type: integer
            format: int32
        name:
          type: string
        nameSpace:
          type: string
        notificationTopicIds:
          type: array
          items:
            type: integer
            format: int32
...
    NotificationTopicDto:
      title: NotificationTopicDto
      type: object
      properties:
        encoding:
          type: string
          enum:
          - JSON
          - AVRO
          - PROTOBUF
        fileFormatIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        messageBus:
          $ref: '#/components/schemas/MessageBusDto'
        messageBusId:
          type: integer
          format: int32
        name:
          type: string
        specificationReference:
          type: string
...